requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pydantic==2.5.0
python-multipart==0.0.6
alembic==1.12.1
celery==5.3.4
redis==5.0.1
docker==6.1.3
PyPDF2==3.0.1
python-docx==1.1.0
regex==2023.10.3
pymongo==4.6.0
openpyxl==3.1.2
google-generativeai==0.3.2
watchdog==3.0.0
numpy<2
huggingface-hub==0.14.1